refactored project structure into Data and Models sub-packages. Initial commit for feature extraction portion of project.
