+++ wsw/data/fingerprint.py
